requirements.txt
Flask==1.0.2
gunicorn==19.9.0
line-bot-sdk==1.7.2
h5py==2.8.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
numpy==1.14.4
pandas==0.20.1
pandocfilters==1.4.2
tensorflow==1.8.0
Pillow==5.1.0